recLog*/requirements.txt
beautifulsoup4
google-generativeai
python-dotenv
requests